requirements.txt
PyQt5==5.15.10
python-vlc==3.0.18121
Pillow==10.1.0
